refactor(allegro,web): apply tech-review fixes for #449 attachments
Surgical follow-ups to the just-shipped #449 attachments UI. No
behaviour change; structural cleanup only.

- Drop the duplicate SafetyAttachmentUploadInput / Result declarations
  in libs/integrations/allegro/src/domain/types/allegro-safety-
  attachments.types.ts; re-export the canonical definitions from
  @openlinker/core/listings instead. Eliminates the contract-drift
  risk where the integrations-local copy could silently strip a new
  port-level field.

- Move use-upload-safety-attachment-mutation.ts from
  features/connections/hooks/ to features/allegro/hooks/. The hook
  hits an Allegro endpoint, so it belongs alongside the allegro API
  module rather than coupling the connections feature to it.

- Replace the runtime-built MIME regex in the upload controller with
  a precomputed ALLEGRO_SAFETY_ATTACHMENT_MIME_PATTERN constant. If
  we widen the accepted MIME set later (the PR body flagged this is
  pending sandbox verification), the pattern is one place to update
  and there's no more "did we escape the right metacharacters" risk.

- Drop uploadedAt from UploadSafetyAttachmentResponseDto (and the FE
  AllegroSafetyAttachmentUploadResponse). It was server-local time
  of receipt, not Allegro-authoritative, and the wizard form state
  doesn't read it. Smaller contract.

Quality gate green: pnpm lint, type-check, test (1213 backend + 772
web tests).
