+++ libs/integrations/allegro/src/index.ts
@@ -59,6 +59,7 @@ export {
   ALLEGRO_SAFETY_ATTACHMENT_UPLOAD_PATH,
   ALLEGRO_SAFETY_ATTACHMENT_MAX_BYTES,
   ACCEPTED_SAFETY_ATTACHMENT_MIME_TYPES,
+  ALLEGRO_SAFETY_ATTACHMENT_MIME_PATTERN,
 } from './domain/types/allegro-safety-attachments.types';
 
 // Module
